common: match si_utime with right format specifier
si_utime and si_stime are of type clock_t, for which no format
specifier is defined. When using a specifier, such as %ld, which is
to be used with "long", make sure the argument really is a "long".

+++ common/ECLogger.cpp
@@ -986,7 +986,8 @@ void generic_sigsegv_handler(ECLogger *lpLogger, const char *app_name,
 	ec_log_bt(EC_LOGLEVEL_CRIT, "Backtrace:");
 	ec_log_crit("Signal errno: %s, signal code: %d", strerror(si->si_errno), si->si_code);
 	ec_log_crit("Sender pid: %d, sender uid: %d, si_status: %d", si->si_pid, si->si_uid, si->si_status);
-	ec_log_crit("User time: %ld, system time: %ld, signal value: %d", si->si_utime, si->si_stime, si->si_value.sival_int);
+	ec_log_crit("User time: %ld, system time: %ld, signal value: %d",
+		static_cast<long>(si->si_utime), static_cast<long>(si->si_stime), si->si_value.sival_int);
 	ec_log_crit("Faulting address: %p, affected fd: %d", si->si_addr, si->si_fd);
 	lpLogger->Log(EC_LOGLEVEL_FATAL, "When reporting this traceback, please include Linux distribution name (and version), system architecture and Kopano version.");
 	/* Reset to DFL to avoid recursion */
